Refactor Candidate components
* Move AddClassificationsScanningPage.jsx, FilterCandidateList.jsx, ScanningPageCandidateAnnotations.jsx and ScanningProfilesList.jsx to the `candidate` directory
* Extract Candidate.jsx, CandidateAutoannotations.jsx, CandidateInfo.jsx, CandidatePhotometry.jsx, CandidateThumbnails.jsx and CustomSortToolbar.jsx from existing candidate components
* Extract `getAnnotationValueString` function to a new helpers.js file
* Add a bit of documentation on candidate files
