+++ static/js/utils/helpers.js
@@ -1,0 +1,19 @@
+const getAnnotationValueString = (value) => {
+  let valueString;
+  const valueType = typeof value;
+  switch (valueType) {
+    case "number":
+      valueString = value.toFixed(4);
+      break;
+    case "object":
+      valueString = JSON.stringify(value, null, 2);
+      break;
+    default:
+      valueString = value.toString();
+  }
+  return valueString;
+};
+
+export {
+  getAnnotationValueString,
+}
